streamlit/requirements.txt
absl-py==2.0.0
alabaster==0.7.13
altair==5.1.2
asttokens==2.4.0
astunparse==1.6.3
atomicwrites==1.4.1
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
bleach==6.1.0
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dm-tree==0.1.8
docutils==0.20.1
executing==2.0.0
feature-engine==0.3.0
flatbuffers==23.5.26
fonttools==4.43.1
gast==0.5.4
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.59.3
h5py==3.10.0
idna==3.4
imageio==2.31.6
imagesize==1.4.1
imbalanced-learn==0.11.0
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.4.0
jupyter_core==5.4.0
kaggle==1.5.16
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
MiniSom==2.3.1
missingno==0.5.2
ml-dtypes==0.2.0
more-itertools==10.1.0
namex==0.0.7
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.0
numpydoc==1.6.0
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
patsy==0.5.3
pi==0.1.2
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
pluggy==1.3.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.6
pure-eval==0.2.2
py==1.11.0
pyarrow==13.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydeck==0.8.1b0
pydotplus==2.0.2
Pygments==2.16.1
pyparsing==3.1.1
pytest==4.3.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
smmap==5.0.1
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
stack-data==0.6.3
statsmodels==0.14.0
streamlit==1.29.0
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.2.0
tifffile==2023.9.26
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.6
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.8
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.14.1
xgboost==2.0.1
zipp==3.17.0
